Remove TD3 optimization and add SAC deployment
